pyproject.toml
[project]
name = "auto-commiter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
]
